+++ src/components/login/LoginForm.jsx
@@ -7,7 +7,8 @@ import { loginSchema } from '@/constants/validationSchema';
 import classNames from 'classnames/bind';
 import styles from './LoginForm.module.css';
 import { useNavigate } from 'react-router-dom';
-import { APIService } from '@api/axios';
+
+import { login } from '@hooks/useLoginHook.js';
 
 const cn = classNames.bind(styles);
 
@@ -22,40 +23,8 @@ export default function LoginForm() {
   });
   const navigate = useNavigate();
 
-  /**
-   * 유저 데이터를 서버에 전송하는 함수
-   * @param {Object} userData
-   * @param {string} userData.loginId - 유저 아이디
-   * @param {string} userData.password - 유저 비밀번호
-   * @returns {Promise}
-   */
-  async function onSubmit(userData) {
-    if (errors.loginId?.message || errors.password?.message) {
-      alert('잘못된 이메일 또는 비밀번호를 입력하셨습니다.');
-    } else {
-      try {
-        const fullEmail = userData.loginId === 'test' ? 'test' : `${userData.loginId}@skuniv.ac.kr`;
-
-        const requestData = {
-          loginId: fullEmail,
-          password: userData.password,
-        };
-
-        const response = await APIService.public.post(import.meta.env.VITE_APP_LOGIN, requestData);
-
-        if (response.success === true) {
-          // localStorage에 토큰 저장
-          localStorage.setItem('token', response.accessToken);
-          localStorage.setItem('refreshToken', response.refreshToken);
-          // 홈화면으로 이동
-          navigate('/');
-        } else {
-          alert('잘못된 이메일 또는 비밀번호를 입력하셨습니다.');
-        }
-      } catch {
-        alert('잘못된 이메일 또는 비밀번호를 입력하셨습니다.');
-      }
-    }
+  function onSubmit(userData) {
+    login(userData, errors, navigate);
   }
 
   return (
